feat(db): create events table with indexes
